+++ CHANGELOG.md
@@ -13,6 +13,10 @@ This is the changelog for RedPepper.
 - Add `install-login` tool for setting up a login for the Manager API.
 - Initialize testing framework with `pytest` and `pytest-trio`.
 
+### Changed
+
+- Use `pydantic` for configuration validation.
+
 ### Fixed
 
 - Update various usages of console assets path.